admin: Resolve future-ui contract class names in unit tests
The future-ui styling contract makes emitted class names public
API, but the mapping lived in `.storybook/main.ts` and applied
only in Storybook. Unit tests saw vitest's own hashed class
names instead, so they could not assert the documented public
names, and the mapping had no test — a regression in it would
silently change the public class names everywhere.

- **The mapping moves to a shared, build-excluded location.**
  Both the Storybook and vitest configs use it, and it configures
  the CSS-module compiler at build time rather than running in a
  consumer, so like `cli/` and `storybook/` it stays out of the
  babel and declaration builds and out of the public exports.
- **The vitest config sets `classNameStrategy: "scoped"` for the
  unit tests.** Without it, vitest generates its own class names
  and ignores the shared `generateScopedName`, so tests would see
  hashed names instead of the contract ones.
