+++ packages/admin/admin/.babelrc.build.json
@@ -1,3 +1,10 @@
 {
-    "ignore": ["**/__stories__/**", "**/*.stories.ts", "**/*.stories.tsx", "src/future-ui/cli/**", "src/future-ui/storybook/**"]
+    "ignore": [
+        "**/__stories__/**",
+        "**/*.stories.ts",
+        "**/*.stories.tsx",
+        "src/future-ui/cli/**",
+        "src/future-ui/cssModules/**",
+        "src/future-ui/storybook/**"
+    ]
 }
