+++ packages/admin/admin/tsconfig.build.json
@@ -4,6 +4,7 @@
     },
     "exclude": [
         "src/future-ui/cli",
+        "src/future-ui/cssModules",
         "src/future-ui/storybook",
         "**/*.test.tsx",
         "**/*.test.ts",
