+++ packages/admin/admin/vitest.config.ts
@@ -3,6 +3,8 @@ import { playwright } from "@vitest/browser-playwright";
 import tsconfigPaths from "vite-tsconfig-paths";
 import { defineConfig } from "vitest/config";
 
+import { generateScopedName } from "./src/future-ui/cssModules/generateScopedName";
+
 export default defineConfig({
     test: {
         reporters: ["default", "junit"],
@@ -12,10 +14,22 @@ export default defineConfig({
         projects: [
             {
                 plugins: [tsconfigPaths()],
+                css: {
+                    modules: {
+                        generateScopedName,
+                    },
+                },
                 test: {
                     name: "unit",
                     environment: "jsdom",
                     setupFiles: "./vitest.setup.ts",
+                    css: {
+                        include: [/\.module\.scss$/],
+                        modules: {
+                            // "stable" (the default) would override the vite-level generateScopedName above.
+                            classNameStrategy: "scoped",
+                        },
+                    },
                 },
             },
             {
